🔧 refactor(Dockerfile, package.json, botservice.ts, mm-client.ts, ImagePlugin.ts, PluginBase.ts, vite.config.ts): package updates and code improvements
In Dockerfile, updated the node version from 16 to 20. This allows us to use the latest features and security patches.

In package.json, updated several packages and added "node-fetch". This allows us to use the latest features and improves the stability and performance of the code.

In botservice.ts, mm-client.ts, ImagePlugin.ts, PluginBase.ts, made several improvements to improve the readability and maintainability of the code.