🔧 refactoring: Updated openai-node to v4 and updated environment variable samples
Updated the OpenAI API from v3 to v4. This allows us to take advantage of the new API features and improves the readability and maintainability of the code. Also, added new environment variable samples to the .env-sample file. This makes it easier for users to configure the use of the new API.

🔄 update(botservice.ts): Updated to OpenAI API v4

The OpenAI API has been updated to v4, and changes have been made to accommodate this. This allows the use of the new API features, improving the performance and functionality of the application. Also, unnecessary code has been removed to improve code readability and maintainability.

🔧 refactor(ImagePlugin.ts, types.d.ts): use string literals instead of directly using OpenAI types

Using OpenAI types directly can lead to unexpected type changes due to library updates. Therefore, we used string literals to ensure type stability. Also, in types.d.ts, we imported the entire OpenAI instead of directly importing OpenAI types. This allows us to reference the necessary types directly, improving the readability of the code.